chore(gallery): Move components, make button text grey
